Make glade installation respect configure prefix
The location that glade files would be installed to was being pulled
from pkgconf, but in a way that the prefix of the directories were not
being overwritten correctly with the custom installation prefix that was
set by configure. This meant that glade files would always be installed
to the system, even if a different prefix was set.

This changes that by using pkgconf's support for setting the prefix to
build the correct path instead of the default system one.